Adjust view at space/accommodation/service info page

+++ umpoint/umpoint-application/src/main/java/my/edu/um/umpoint/modules/accommodation/dto/AccAccommodationDTO.java
@@ -62,7 +62,7 @@ public class AccAccommodationDTO implements Serializable {
 	private String facilities;
 
 	@SchemaProperty(name = "Max capacity")
-	private BigDecimal capacity;
+	private Integer capacity;
 
 	@SchemaProperty(name = "Manager ID")
 	private Long manager;
